image-pipeline/model-serving/pyproject.toml
[project]
name = "model-serving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "open-clip-torch>=3.1.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.0",
    "tritonclient[grpc]>=2.59.0",
    "uvicorn>=0.27.0",
]
